backend/requirements.txt
fastapi
uvicorn[standard]
pdfplumber
pandas
python-multipart
openpyxl
google-genai
python-dotenv
playwright
